move select dinner state to mainView
